Moved Console Printer class into a new package under utils.
MOdified the way this class works so that you can add data from anywhere in our application. No more need to expose every subsystems sensors publically just so data can be logged. This also avoids any possibilities of the console printer attempting to access uninitialized classes, as the classes themselves add data to the console printer for logging. Data onlt needs to be added once, and thereafter it will be periodically captured at the specified rate. A convenient place to initiate this logging is within the constructor of each subsystem.

I'm using a new feature in Java 8 (lambda expressions). This is effectively providing function pointers to the methods within each subsystem that can provide the data that needs to be logged. The lambda expressions are passed in to the Console Printer. Later, after the logging thread is started, it periodically calls the lambdas; providing the latests values for each data element.

You no longer need to babysit the formatting of the logged text file either. The tabbed delimited format file is auto generated from the set of enrolled data elements. I'm using Linked hashmaps for the keys, so the order the data is logged will be deterministic; it will match the order in which each data element is added to the set for logging. New elements added for logging will be added to the end of the list.