Moved command into subsystem specific package.
Renamed local variables for motors to match naming convention used in robot map. left/right instead of 1/2.

The order in which we reverse the motor polarity and check for positive only values is important. We want calls into this method to be able to pass in values that are positive regardless of what's being sent out to the motor controllers. Because of the way they are mounted (facing away from one another), one motor will always need to be turned in a direction opposite of the other. So one motor will need negative values sent.

If we check for a need to reverse the motor inputs before we check for negative values, one motor will always fail the check. Need to check for values less than zero before the motor inversion is applied.